pyproject.toml
[project]
name = "c2a-pytest-gaia"
version = "0.1.0"
description = ""
authors = []

requires-python = ">= 3.9"
dependencies = [
    "grpcio~=1.54.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black~=23.1.0",
    "grpcio-tools~=1.54.2",
]
[tool.hatch.metadata]
allow-direct-references = true
